[Decomp][Unit] Cohesion split: extract visibility/detection to UnitVisibility.cpp
